+++ src/soromox/systems/planar_pcs.py
@@ -474,7 +474,7 @@ class PlanarPCS(DynamicalSystem):
         """
 
         # Classify the point along the robot to the corresponding segment
-        segment_idx = jnp.clip(jnp.sum(s > self.L_cum) - 1, 0, self.num_segments - 1)
+        segment_idx = jnp.clip(jnp.sum(s > self.L_cum) - 1, 0, self.num_segments - 1).astype(jnp.int32)
 
         # Compute the point coordinate along the segment in the interval [0, l_segment]
         s_local = s - self.L_cum[segment_idx]
@@ -592,7 +592,6 @@ class PlanarPCS(DynamicalSystem):
         xi = self.strain(q).reshape(self.num_segments, 3)
 
         segment_idx, s_local = self.classify_segment(s)
-        segment_idx = segment_idx.astype(jnp.int32)
         s_local = jnp.asarray(s_local, dtype=xi.dtype)
 
         zeros = jnp.zeros((self.num_segments, 3, 3), dtype=xi.dtype)
@@ -794,7 +793,6 @@ class PlanarPCS(DynamicalSystem):
         xid = (self.B_xi @ qd).reshape(self.num_segments, 3)
 
         segment_idx, s_local = self.classify_segment(s)
-        segment_idx = segment_idx.astype(jnp.int32)
         s_local = jnp.asarray(s_local, dtype=xi.dtype)
 
         zeros = jnp.zeros((self.num_segments, 3, 3), dtype=xi.dtype)
